+++ scores/trashfire/_tests/trash-actions.js
@@ -1,20 +1,7 @@
 const path = require('path')
 const { loadDomThenTest } = require(path.resolve('.', 'bin/js/tape-setup'))
 
-loadDomThenTest.only('trashfire', '_site/scores/trashfire/index.html', (t, window) => {
-    // const trashActions = ['add', 'remove', 'add', 'add', 'remove', 'add', 'add', 'add', 'empty']
-
-    const { last, push } = window.TrashUtils
-
-    const abcd = ['a', 'b', 'c', 'd']
-    let d = last(abcd)
-    t.equals(d, 'd', 'last function should return last element of array')
-    d = 'e'
-    t.deepEquals(abcd, ['a', 'b', 'c', 'd'], 'last function should not mutate source array')
-
-    const abc = ['a', 'b', 'c']
-    t.deepEquals(push(abc, 'd'), ['a', 'b', 'c', 'd'], 'push function should return a new array with an added item')
-    t.deepEquals(abc, ['a', 'b', 'c'], 'push function should not mutate source array')
+loadDomThenTest('trashfire', '_site/scores/trashfire/index.html', (t, window) => {
 
     const { addTrash, removeTrash, emptyTrash, copyTrash } = window
